Update schema for routing protocol and remove fabric_mapping_helper file by moving contents to correct resource files
